requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==2.2
django-environ==0.4.5
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
toml==0.10.2
whitenoise==5.2.0
django-storages
google-cloud-storage